Introduction Bussiness Problem Car Accident Severity
IBM Data Science Capstone Report
Car Accident Severity 
in Seattle
Introduction/Business Problem
Car accidents are one of the most common hazards we all face in daily life. From minor accidents to catastrophic, multi-car pileups, getting into an accident in a motor vehicle can end a life or change it forever. 
Car accidents happen for  many of reasons, including behavioral, environmental, and situational. A small number of car accidents are inevitable and can’t be prevented. Most of them, however, could at least be prevented, and many result from poor decisions by someone who should have done better. Among the most common human-related causes of car accidents are Alcohol and drug, drivers distraction and speeding.
Seattle is a big city and when it comes to car accidents in Seattle, the city is certainly not the worst in the United States, but we aren’t the safest either. The  foggy roads, occasional black ice, and frequent rain showers present hazards that we can’t control. But what about those aspects that you, as a driver, CAN control?
The  most common causes of car accidents are:
Distracted Driving: driving while scrolling down an app feed on their phone, calling, texting or reading..
Driving Under the Influence: many terrible accidents have been caused by cocaine, heroin, marijuana, prescription pills and other drugs too
Speeding
 Reckless Driving: driving without respecting the rules, the stop sign...
Weather conditions: heavy rain, heavy snow....
According to The Seattle Department of Transportation, in  2017, Seattle police reported 10,959 motor vehicle collisions on city streets. According to the report, in 2017 there were a total of 187 fatal and serious injury collisions on Seattle streets. Data available from the Washington State Department of Transportation (WSDOT) reflect an even worse tally in 2018, with 212 crashes that resulted in serious injury or wrongful death. 
In an effort to reduce the frequency of car collisions in Seattle, this project aim to predict   the severity of an accident knowing  the current weather, road and visibility conditions,  the speeding, and whether the driver was uninfluenced of drugs or alcohol and to alert the drivers to be more careful on the  road when condition are not When conditions are not reunite, this model will alert drivers  to be more careful.